feat: defined the domain models with jpa (#3)
* feat: created user-related data model classes

I have created the POJO's and the required enums for the user-related classes with all necessary jpa annotations. Hibernate should now be able to map them to database tables with their relations

* feat: created event- and chat-related data model classes

I have created the POJO's and the required enum for the chat- and event-related classes with all necessary jpa annotations. Hibernate should now be able to map them to database tables with their relations

* feat: created matching-related data model classes

I have created the POJO's and the required enum for the matching-related classes with all necessary jpa annotations. Hibernate should now be able to map them to database tables with their relations

* feat: added field for picture to User and MatchLevel entity

I have added a BLOB field in the User and MatchLevel entity classes, so the users profile picture and the respective match level pictures can be persisted in the database

* feat: added null validation to not nullable entity fields

I have added a null validation to the entity fields, so the fields can't be set to null accidentally

* feat: added many to many mapping in user entity and lazy fetch type to conversation entity

I have added lazy fetching to the participants field in the conversation entity to increase the performance. I have also included the participatingConversations field in the user entity, so I can directly access the participating conversations in the user model