feat(api): declared endpoints and schemas for managing users (#8)
* fix(model): removed requirement for a user preference in the user model

I have removed the requirement for a user preference. A user preference is only necessary if the user wants to use the dating feature, and should therefore only be saved in the database if the user has enabled the dating feature in his settings.

* chore: excluded the gen directory from git vcs

The gen directory contains all generated files from the open api generator. It contains a lot of unnecessary files which should not be tracked by git. Therefore, I have excluded the whole directory from the git vcs

* feat(api): defined the user dto in the api specification

I have defined a common user dto, so every api consumer can see, what a user object should look like and which properties a user has. Due to changed customer opinion, I have also made the profession and the department of a user mandatory

* fix(model): fixed typo in 'proposed'

* fix(api): fixed wrong regex pattern

The old regex pattern didn't allow german umlauts or hyphens in the name like in karl-heinz

* chore: moved enums to an own package

The enums were mixed in the same package with the entities. Moving them to an own package should make the project structure much cleaner

* refactor(model): removed the department model & included it into the user model

Because the department only had one field, it was unnecessary to save it as an own entity. Therefore, I have included it as a string field into the user model

* perf(model): set user preferences' fetch type to lazy in the user model

The preferences of a user aren't needed until they are requested specifically. Therefore, setting the fetch type to lazy can improve the performance of the application

* refactor(api): included user tag and renamed api base path to api version

'Api version' seems more suitable for describing the actual version of the api instead of 'base path'. I have also introduced a user tag, so all operations related to the user are grouped in the api documentation

* feat(api): declared paths for basic CRUD operations of a user

I have declared the paths for creating, querying, updating and deleting a user in the api specification. I have also renamed the user_preference_id column to user_preferences_id, because it felt more suitable

* feat(api): introduced a query limit for the 'get all users' endpoint

I have introduced a query parameter which allows to specify, how many user records should be queried from the database, so the administrator using the frontend application can control, how many user records should be loaded

* refactor: made the api descriptions more verbose

Some descriptions were not clear enough, so I replaced them with a more suitable and verbose one. I have also renamed the 'query limit' parameter to 'max_results', which seems more suitable

* fix(model): made specification of the department required in the user model

The product owner wants the specification of the users department to be necessary. I have introduced a null validator to ensure that only non-null values are stored to the database