feat: implemented user registration (#10)
* build(gradle): installed and configured the open api generator plugin

In order to generate the api interfaces automatically, I have implemented and configured a gradle task, which runs the open api generator every time the project compiles. In order to get access to the generated data from the main class path, I have also created a source set which links the two class paths

* build(gradle): open api generator appends automatically the 'Dto' suffix to DTOs

The models and the DTOs had the same filename, so I have configured the open api generate task to append 'Dto' to every specified schema in the specification.

* build(gradle): added missing dependencies for the open api generator

The open api tool requires the spring validation and the swagger core dependency, which were missing. I have also added some options to the open api configuration

* fix(model): changed id data type from long to int & moved user model to own directory

* fix: introduced custom base path in application-dev.properties

The clients call the endpoints with localhost:port/api/v1 as base url. The openapi generator didn't generate the custom base path, therefore it has to be introduced in the application-dev.properties

* chore(build): changed the base package of the generated api to de.icp.match

* feat: implemented user creation functionality

* refactor(api): split UserDTO into UserDTO and UserCreationDTO

The open api generator isn't capable of generating read- or writeOnly properties specified in the oas. Therefore, I had to spit up the UserDTO into a DTO only for querying operations, and one for creating and manipulating operations

* build: introduced library to automatically generate mapper implementations

I have included the MapStruct library, which is capable of automatically generating the implementations for the mapper methods defined in the respective mapper interface

* refactor: adapted the user controller to the new UserCreationDto

* test(user service): created unit tests for creating unique usernames

* test(user service): extended unit tests for creating users

I have created a unit test for checking if the password gets encrypted (not implemented yet). I have also renamed the tests to more comprehensive names.

* build(deps): included a h2 database for running tests

Integration tests won't succeed if they run on a test system, because they don't have access to the database. The h2 database is an in-memory db, so it allows to perform database actions independently of the underlying system.
Because there is already a user table in the h2 db, I had to dp a workaround through 'renaming' the user table of the application

* test(user controller): created integration tests for creating a user

I also removed the service annotation from the UserCreationService, because the implementation had it too, and spring didn't know from which interface/class it should create a bean

* feat(user creation): added password encryption and basic security configuration

In order to provide a password encoder in the userCreationService, I had to create the SecurityConfig, which also disables cors and the csrf protection, which won't be needed when using jwt for authentication

* test(user creation): added test for checking the password encryption